requirements.txt
beautifulsoup4         4.11.1
brotlipy               0.7.0
catalogue              2.0.6
certifi                2021.10.8
cffi                   1.15.0
charset-normalizer     2.0.12
ChatterBot             1.0.8
chatterbot-corpus      1.2.0
click                  7.1.2
colorama               0.4.4
conda                  4.12.0
conda-package-handling 1.8.1
cryptography           36.0.2
cymem                  2.0.6
cytoolz                0.9.0.1
dataclasses            0.8
decorator              5.1.1
dill                   0.2.9
en-core-web-sm         2.0.0
future                 0.18.2
gTTS                   2.2.4
gTTS-token             1.1.4
idna                   3.3
Jinja2                 3.1.2
langcodes              3.3.0
markovify              0.9.4
MarkupSafe             2.1.1
mathparse              0.1.2
menuinst               1.4.18
mkl-fft                1.0.6
mkl-random             1.0.1
msgpack                0.6.1
msgpack-numpy          0.4.3.2
murmurhash             1.0.7
numpy                  1.15.4
opencv-python          4.2.0.32
packaging              21.3
pathy                  0.6.1
Pillow                 9.0.1
pip                    21.2.4
plac                   0.9.6
preshed                2.0.1
pycosat                0.6.3
pycparser              2.21
pydantic               1.8.2
pyOpenSSL              22.0.0
pyparsing              3.0.9
pyreadline             2.1
PySocks                1.7.1
python-dateutil        2.8.2
python-decouple        3.6
python-telegram-bot    12.4.2
pytz                   2022.1
pywin32                303
PyYAML                 3.13
regex                  2022.3.15
requests               2.27.1
ruamel-yaml-conda      0.15.80
setuptools             61.2.0
shellingham            1.4.0
six                    1.16.0
smart-open             5.2.1
soupsieve              2.3.2.post1
spacy                  2.0.16
spacy-legacy           3.0.9
spacy-loggers          1.0.2
SQLAlchemy             1.3.24
srsly                  2.4.3
TBB                    0.2
thinc                  6.12.1
toolz                  0.11.2
tornado                6.1
tqdm                   4.64.0
typer                  0.4.1
typing-extensions      3.10.0.2
ujson                  5.1.0
Unidecode              1.3.4
urllib3                1.26.9
wasabi                 0.9.1
wheel                  0.37.1
wikipedia              1.4.0
win-inet-pton          1.1.0
wincertstore           0.2
wrapt                  1.10.11
zipp                   3.8.0
